// src/api/types.gen.ts
export type ClientSettingsFolderRequest = {
    name: string;
    moveHereWhen?: string;
    friendPublicIds?: Array<number>;
};

export type UpdateClientSettingsRequest = {
    resolution?: string;
    uiScale?: number;
    accentColor?: string;
    background?: string;
    clientAnimation?: string;
    language?: string;
    chatPosition?: string;
    screenMode?: string;
    allowFriendRequest?: string;
    showEmailPublic?: boolean;
    useFriendColors?: boolean;
    folders?: Array<ClientSettingsFolderRequest>;
};

export type ClientSettingsFolderResponse = {
    name?: string;
    moveHereWhen?: string;
    friendPublicIds?: Array<number>;
};

export type ClientSettingsResponse = {
    resolution?: string;
    uiScale?: number;
    accentColor?: string;
    background?: string;
    clientAnimation?: string;
    language?: string;
    chatPosition?: string;
    screenMode?: string;
    allowFriendRequest?: string;
    showEmailPublic?: boolean;
    useFriendColors?: boolean;
    folders?: Array<ClientSettingsFolderResponse>;
};